Refactored models to go into its own namespace.
