+++ GraduationTracker/GraduationTracker/Models/Course.cs
@@ -1,0 +1,10 @@
+﻿namespace GraduationTracker.Models
+{
+    public class Course
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public int Mark { get; set; }
+        public int Credits { get; }
+    }
+}
