+++ GraduationTracker/GraduationTracker/Models/Diploma.cs
@@ -1,0 +1,9 @@
+﻿namespace GraduationTracker.Models
+{
+    public class Diploma
+    {
+        public int Id { get; set; }
+        public int Credits { get; set; }
+        public int[] Requirements { get; set; }
+    }
+}
