+++ GraduationTracker/GraduationTracker/Models/Student.cs
@@ -1,0 +1,9 @@
+﻿namespace GraduationTracker.Models
+{
+    public class Student
+    {
+        public int Id { get; set; }
+        public Course[] Courses { get; set; }
+        public STANDING Standing { get; set; } = STANDING.None;
+    }
+}
